Move song related helpers to songInfo.ts
